requirements.txt
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-decouple==3.8
pytz==2023.3.post1
